fraud-suite/requirements.txt
flask
flask-socketio
nfcpy
pycryptodome
transformers
